Move test to appropriate location
